+++ backend/expense/src/main/java/org/egov/digit/expense/kafka/ExpenseProducer.java
@@ -1,22 +1,36 @@
 package org.egov.digit.expense.kafka;
 
+import lombok.extern.slf4j.Slf4j;
+import org.egov.common.utils.MultiStateInstanceUtil;
 import org.egov.tracer.kafka.CustomKafkaTemplate;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 // NOTE: If tracer is disabled change CustomKafkaTemplate to KafkaTemplate in autowiring
 
+@Slf4j
 @Service
 public class ExpenseProducer {
 
     private final CustomKafkaTemplate<String, Object> kafkaTemplate;
+    private final MultiStateInstanceUtil multiStateInstanceUtil;
 
     @Autowired
-    public ExpenseProducer(CustomKafkaTemplate<String, Object> kafkaTemplate) {
+    public ExpenseProducer(CustomKafkaTemplate<String, Object> kafkaTemplate, MultiStateInstanceUtil multiStateInstanceUtil) {
         this.kafkaTemplate = kafkaTemplate;
+        this.multiStateInstanceUtil = multiStateInstanceUtil;
     }
 
-    public void push(String topic, Object value) {
-        kafkaTemplate.send(topic, value);
+    /**
+     * Publishes a message to a Kafka topic based on the tenant-specific topic name.
+     *
+     * @param tenantId the unique identifier of the tenant for which the message is being published
+     * @param topic the base Kafka topic name to which the message needs to be pushed
+     * @param value the message payload to be sent to the Kafka topic
+     */
+    public void push(String tenantId, String topic, Object value) {
+        String updatedTopic = multiStateInstanceUtil.getStateSpecificTopicName(tenantId, topic);
+        log.info("The Kafka topic for the tenantId : {} is : {}", tenantId, updatedTopic);
+        this.kafkaTemplate.send(updatedTopic, value);
     }
 }